Fix: Corrige o bug realcionado ao momento que o usuário confirma se ele aceita o aluguel
